requirement.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.6
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-two-factor-auth==1.17.0
pillow==11.3.0
pypng==0.20220715.0
python-dotenv==1.1.1
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
